requirements.txt
nanogps==0.1.0
python-dotenv==0.19.1
opencv-python==4.5.4.58
scikit-learn==0.24.2
python-telegram-bot==13.7

